refactoring to user real map tiles stored on github
